Add files via upload
This project will use the Structured Query Language (SQL) to explore and analyze data related to the COVID-19 pandemic. The project will answer a variety of questions about the pandemic, such as how many people have been infected with COVID-19, how many people have died from COVID-19, what are the trends in COVID-19 cases and deaths over time, what are the factors that contribute to the spread of COVID-19, and what are the most effective measures for preventing the spread of COVID-19.